Changes to schedulers, new default scheduler added
+ Added new sequential scheduller that is now used by default, which
handles cache lines much better thus increases performance.
* Added some comments to the schedulers that exist.
* Moved the Add and Remove methods from the scheduler interface to a new
interface. The reseon for this is that those are only needed when
accesing the artificial scheduler (when someone creates a new thread and
not a computation)  other scheulers that work on tread/fiber pools don't
use those methods and potential synch of those methods is complex and
time consuming.